# docs/swagger.json
{"openapi":"3.0.0","paths":{"/api/service-manager-service/v1":{"get":{"operationId":"AppController_find","parameters":[],"responses":{"200":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDefault"}}}}},"tags":["Generic"]}},"/api/service-manager-service/v1/users/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/service-manager-service/v1/users":{"get":{"operationId":"UsersController_getUsers","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}]}}}}},"tags":["Users"]},"post":{"operationId":"UsersController_signupUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/api/service-manager-service/v1/requirements/{id}":{"get":{"operationId":"RequirementsController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["requirements"]},"put":{"operationId":"RequirementsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequirementsDto"}}}},"responses":{"200":{"description":""}},"tags":["requirements"]},"delete":{"operationId":"RequirementsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["requirements"]}},"/api/service-manager-service/v1/requirements":{"get":{"operationId":"RequirementsController_getUsers","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RequirementsEntity"}}}}]}}}}},"tags":["requirements"]},"post":{"operationId":"RequirementsController_signupUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequirementsDto"}}}},"responses":{"201":{"description":""}},"tags":["requirements"]}}},"info":{"title":"Generic Adapter ","description":"Adapter of Generic backend repo","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ErrorDefault":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"},"error":{"type":"string"}},"required":["statusCode","message","error"]},"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","firstName","lastName","password","createdAt","updatedAt"]},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"RequirementsEntity":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"userId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","userId","createdAt","updatedAt"]},"CreateRequirementsDto":{"type":"object","properties":{"userId":{"type":"number"},"title":{"type":"string"}},"required":["userId","title"]},"UpdateRequirementsDto":{"type":"object","properties":{"userId":{"type":"number"},"title":{"type":"string"}}}}}}